Store hash of feed URL as well as UID.
This is so we can ensure we only delete events which came from the feed we're
processing now - so if this script is being used to sync multiple iCal feeds
into the same Google Calendar, we delete only imported events which came from
the feed currently being processed, ratehr than each execution deleting all
imported events from all other feeds.

This is for Issue #8.