Start of tests using mocking
A lot more to do, but committing this now.

Currently prepares a mocked Net::Google::Calendar object, passes it to the
gcal_obj class method, and tests that we then get that mocked object back when
we ask for a Net::Google::Calendar object.

Next up: tests that, when given known iCal data, the results stored in the mock
object are as we'd expect.