For each ical file, check results look sane.
Not yet complete - more hacking to do here - but handles looping over a set of
test specs, each of which names an iCal data file to parse, and provides a
hashref describing what the resulting entries added to the Google calendar ought
to look like.

Will need a flag in each test spec to indicate if the entries in the mock gcal
object should be wiped before starting, so some tests can be testing only that
the results from that parse are correct, but others can test that the results
look as expected after making changes based on the previous one too (to e.g.
test for events that are no longer in the ical feed / have been modified etc
being handled correctly).