Don't change time when setting timezone.
We want to return DateTime objects using the timezone from the iCal feed, but we
don't want to call set_time_zone() to set the timezone as I was, because that
changes the time represented by the object too.

This was is a bit clumsy but works.

FIXME: it seems to be dog-slow - probably because it's calculating loads of
DST changes etc.  May have to revert this and implement my own parsing of the
ISO8601 datetimes, and just create DT objects with the right datetime and
timezone set?  Or maybe Data::ICal has something useful.